Revert "Add reviews"

+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/ReviewServiceImpl.java
@@ -46,20 +46,22 @@ public class ReviewServiceImpl implements ReviewService {
     @Override
     public ReviewDto updateReview(UpdateReviewRequest request) {
         Review review = reviewRepository.getReviewById(request.getReviewId());
-
+        if (review == null) {
+            throw new NotFoundException("Отзыв не найден");
+        }
 
         ReviewMapper.updateReview(review, request);
         reviewRepository.updateReview(review);
-
-
         log.info("ReviewServiceImpl: отзыв с id {} обновлен", review.getReviewId());
         return ReviewMapper.mapToReviewDto(review);
     }
 
     @Override
     public void deleteReview(int reviewId) {
-
-
+        Review review = reviewRepository.getReviewById(reviewId);
+        if (review == null) {
+            throw new NotFoundException("Отзыв не найден");
+        }
         reviewRepository.deleteReview(reviewId);
         log.info("ReviewServiceImpl: отзыв с id {} удален", reviewId);
     }
@@ -95,36 +97,31 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
     @Override
-
-    public ReviewDto addLikeReview(int reviewId, int userId) {
+    public void addLikeReview(int reviewId, int userId) {
+        validateReviewAndUser(reviewId, userId);
         reviewRepository.addLikeReview(reviewId, userId);
         log.info("Пользователь {} лайкнул отзыв {}", userId, reviewId);
-        return null;
     }
 
     @Override
-    public ReviewDto addDislikeReview(int reviewId, int userId) {
+    public void addDislikeReview(int reviewId, int userId) {
         validateReviewAndUser(reviewId, userId);
         reviewRepository.addDislikeReview(reviewId, userId);
         log.info("Пользователь {} дизлайкнул отзыв {}", userId, reviewId);
-        return null;
     }
 
     @Override
-    public ReviewDto deleteLikeReview(int reviewId, int userId) {
+    public void deleteLikeReview(int reviewId, int userId) {
         validateReviewAndUser(reviewId, userId);
         reviewRepository.deleteLikeReview(reviewId, userId);
         log.info("Пользователь {} убрал лайк с отзыва {}", userId, reviewId);
-        return null;
     }
 
     @Override
-    public ReviewDto deleteDislikeReview(int reviewId, int userId) {
+    public void deleteDislikeReview(int reviewId, int userId) {
         validateReviewAndUser(reviewId, userId);
         reviewRepository.deleteDislikeReview(reviewId, userId);
         log.info("Пользователь {} убрал дизлайк с отзыва {}", userId, reviewId);
-        return null;
-
     }
 
     @Override
@@ -135,4 +132,3 @@ public class ReviewServiceImpl implements ReviewService {
     }
 
 }
-
